Move formatting functions to single file
