+++ .github/workflows/build-test.yml
@@ -52,3 +52,6 @@ jobs:
 
       - name: Lint
         run: pnpm lint
+
+      - name: Typecheck
+        run: pnpm tsc
