--- src/utils/format-currency.ts
@@ -1,6 +1,0 @@
-export function formatCurrency(amount: number, currency: string, locale: string = "en-US") {
-  return new Intl.NumberFormat(locale, {
-    style: "currency",
-    currency,
-  }).format(amount);
-}
